rename: `getLoginUri` → `useLoginUri`로 변경
[#22]
